Done and Dusted Order and Item management

+++ src/lk/nnj/rms/fx/service/IAdminQueryService.java
@@ -7,8 +7,8 @@ public interface IAdminQueryService {
     double findTotalSales(LocalDate d1, LocalDate d2) throws Exception;
     int findTotalItems(LocalDate d1, LocalDate d2) throws Exception;
     int findTotalCustomers(LocalDate d1, LocalDate d2) throws Exception;
-    int findTotalSalesPerDay(LocalDate d1) throws Exception;
-    int findDineInSalesPerDay(LocalDate d1) throws Exception;
-    int findTakeAwaySalesPerDay(LocalDate d1) throws Exception;
-    int findDeliverSalesPerDay(LocalDate d1) throws Exception;
+    double findTotalSalesPerDay(LocalDate d1) throws Exception;
+    double findDineInSalesPerDay(LocalDate d1) throws Exception;
+    double findTakeAwaySalesPerDay(LocalDate d1) throws Exception;
+    double findDeliverSalesPerDay(LocalDate d1) throws Exception;
 }
